Get rid of ArticleContext
